Rename klplib/patch to klplib/analysis
`patch` name is too ambiguous nowadays. On the other hand,
`analysis` is the common way of referencing the file within
internal documentation and between mantainers.
